Reverting a warning fix that collided with a fix res committed...
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23035 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/evoutlet.cpp
@@ -97,7 +97,7 @@ void csEventOutlet::Mouse (uint iNumber, int iButton, bool iDown, const int32 *a
   iMouseDriver* m = GetMouseDriver();
   if (m != 0)
   {
-    if (iButton == (uint) csmbNone)
+    if (iButton == csmbNone)
       m->DoMotion (iNumber, axes, numAxes);
     else
       m->DoButton (iNumber, iButton, iDown, axes, numAxes);
